requirements.txt
ninja
moviepy==1.0.3
numpy==1.24.3
scipy==1.14.1
matplotlib==3.9.2
pillow==10.4.0
opencv-python==4.10.0.84
einops==0.8.0
timm==1.0.10
torch-scatter==2.1.2
hydra-core==1.3.2
omegaconf==2.3.0
tqdm==4.66.5
imageio==2.36.0
packaging
evo==1.30.1
pandas==2.2.3
seaborn==0.13.2
rerun-sdk==0.19.0
ipython==8.28.0
pytest==8.3.3
tensorboard==2.18.0
pypose==0.6.8